refactor: split Paragon Plugin Utils file
